fix: recognize conditional-expression and unpacked-loop Fact aliases
Two independent Codex-review findings on the fact-detector-misuse gate:

1. _is_fact_typed_expr() never unwrapped a conditional expression -
   `(old.bases_fact if condition else new.bases_fact) == other` (and the
   equivalent assign-then-compare form) both genuinely produce a
   Fact-typed result regardless of which branch runs, but ast.IfExp had
   no branch here. Fixed by requiring both node.body and node.orelse to
   independently resolve as Fact-typed - deliberately narrower than the
   existing NamedExpr unconditional unwrap, since only the case where
   both branches are guaranteed Fact-typed is a real unconditional
   result.

2. The loop-target literal-collection fix only ever handled a bare
   ast.Name target - a tuple-unpacking target (`for fact, tag in
   ((old.bases_fact, "old"), (new.bases_fact, "new")):`) was silently
   excluded. Fixed with a new branch reusing
   _paired_unpacking_candidates() once per iteration element, requiring
   every iteration to pair successfully (any failure disqualifies the
   whole loop) before applying the same conjunctive
   every-element-Fact-typed-or-deferred-name requirement.

Both verified against real repros plus negative controls. Zero existing
hits; mypy/ruff clean.

Also: split scripts/fact_detector_misuse.py into a sibling leaf module,
scripts/fact_detector_misuse_scope.py, once these fixes pushed the file
past the AI-readiness file-size gate's 2000-line hard cap. Mechanical
extraction (unchanged function bodies) of the eight lexical-scope-
resolution building blocks the rest of the module builds its alias
resolution on. Verified both isolated test files (which import
scripts.fact_detector_misuse directly, without importing
check_ai_readiness first) still pass on their own, not just as part of
the full suite - exactly the scenario a missing sys.path guard would
silently pass in a full run and fail only in isolation. Registered in
scripts/CLAUDE.md's inventory table.

New tests: TestConditionalExpressionsRecognizedWhenBothBranchesAreFactTyped
and TestForLoopUnpackingTargetsResolveElementwise in
tests/test_fact_detector_misuse_def_time_scope.py.
