+++ scripts/CLAUDE.md
@@ -16,6 +16,7 @@ Each must run with Python 3.10+ and the package installed in dev mode
 | `adr_status_sync.py` | The `adr-status-sync` gate, split out of `check_ai_readiness.py` (already past the 2000-line hard cap): an ADR's own `**Status:**` line and its row in `docs/contribute/adr/index.md` may not *contradict* each other, an optional `**Verified:** <ref>@<sha> on <date>` receipt must parse, and commits after that sha touching a first-party file the Status names raise a WARN — the only check here that catches *document-vs-code* drift. Also owns the ADR Status parsing helpers (`_ADR_FILE_RE`, `_adr_status_text`) that `check_ai_readiness.py`'s own `adr-index-nav-sync` reuses. Pure-stdlib, importable. | imported (by `check_ai_readiness.py`); gated via `tests/test_ai_readiness.py`. |
 | `engine_cli_boundary.py` | The `engine-cli-boundary` gate, split out of `check_ai_readiness.py` the same way `adr_status_sync.py` was (same 2000-line-cap reason): no engine-layer module (`scan_engine.py`, `service*.py`, `artifact_*.py`, `buildsource/**/*.py`, `workflows/artifact/**/*.py`) may import `click` or a `cli`/`cli_*` sibling — including a *nested* CLI adapter like `abicheck/compat/cli.py` — allowlist-and-shrink (`ENGINE_CLI_BOUNDARY_ALLOWLIST`), keyed by a stable `<rel>::<desc>::<occurrence>` identity rather than line number. A trailing imported alias is only flagged once verified against the real file tree as an actual submodule (`_alias_is_real_submodule`), so an ordinary symbol that merely starts with `cli_` is never a false positive. Pure-stdlib, importable. | imported (by `check_ai_readiness.py`); gated via `tests/test_engine_cli_boundary.py`. |
 | `fact_detector_misuse.py` | The `fact-detector-misuse` gate (ADR-063 Phase 0, `docs/contribute/plans/one-semantic-pipeline.md`), split out of `check_ai_readiness.py` the same way `fact_field_readers.py`/`engine_cli_boundary.py` were: a real, repo-wide AST scan for a direct `==`/`!=` comparison of a `Fact[T]`-typed value (a `<attr>_fact` field access, or a `Fact(...)`/`Fact.<classmethod>(...)` constructor call) — the check `abicheck/model/fact.py`'s own `Fact` docstring already claimed exists by name before this module was written. No baseline: zero existing hits under `abicheck/` today, so any match is an unconditional error. Pure stdlib, importable. | imported (by `check_ai_readiness.py`); gated via `tests/test_fact_detector_misuse.py`. |
+| `fact_detector_misuse_scope.py` | The lexical-scope resolution primitives `fact_detector_misuse.py` builds its own alias resolution on — `_enclosing_qualnames`/`_qualname_at`/`_QualnameSpans`, `_lexical_function_parents`, `_def_containing_qualnames`, `_bound_names`, `_paired_unpacking_candidates`, `_match_pattern_names` — split out once `fact_detector_misuse.py` itself crossed the AI-readiness `file-size` gate's own 2000-line hard cap; a mechanical extraction (unchanged function bodies), not a redesign. No public entry point of its own — imported by `fact_detector_misuse.py` only, via a sys.path guard mirroring `check_ai_readiness.py`'s own identical one (so it resolves whether `fact_detector_misuse.py` runs directly or is loaded as `scripts.fact_detector_misuse` by a test that never imported `check_ai_readiness.py` first). Pure stdlib, importable. | imported (by `fact_detector_misuse.py`); gated transitively via `tests/test_fact_detector_misuse.py`/`tests/test_fact_detector_misuse_scoping.py`/`tests/test_fact_detector_misuse_def_time_scope.py`. |
 | `findings_report.py` | The shared `Findings` error/warning collector both gate scripts report through — grouping by check name, the `ERROR:`/`WARN:` rendering, and the `<label>: N error(s), M warning(s)` summary line. `check_ai_readiness.py` and `check_docs_contract.py` each subclass it to pin their own label; previously each carried a byte-identical copy of the class. Pure-stdlib (importable before `pip install -e .`, same constraint as `check_ai_readiness.py` itself). | imported (by `check_ai_readiness.py`, `check_docs_contract.py`). |
 | `check_fp_rate.py` | False-positive/false-negative gate for public-surface scoping (ADR-024 §7). Labelled `(old, new)` corpus; baselines FP=0/FN=0. Cases are tagged by scoping *axis* (`CASE_CATEGORY`): `--json` carries a `by_category` breakdown and `--markdown` renders a per-axis table for CI step-summary / release-over-release trend reading. | CI (`ai-readiness`). Mirrored in `tests/test_fp_rate_gate.py`. |
 | `check_tier_accuracy.py` | Per-evidence-tier accuracy gate — *what each level buys*. One labelled logical change per case, projected down to what each tier observes (L0 symbols → L1 debug → L2 headers → L3 build) and run through `compare`; verdicts collapse to a 3-band ordinal (non-breaking/risk/breaking). Measures per-tier **over-call (FP)** vs **under-call (FN)** and which transition removes each. Gates: top-tier correctness + under-call monotonicity (more evidence never hides a break — ADR-028 D3). `--markdown`/`--json` emit the per-tier matrix. | CI (`ai-readiness`, + step-summary artifact). Mirrored in `tests/test_tier_accuracy_gate.py`. |
